adding pythonConvert script to prebuild
